[Tests] Merge the five telemetry test files into one, and drop the duplication
tests/telemetry held five files, and four of them carried their own copy of
the same scaffolding: the in-memory exporter fixture, the `spans` fixture,
`FakeLLM`, `fake_agent`, and the span helpers. Merged into a single file with
each defined once.

  5 files, 3,361 lines, 162 tests, 3 failing
  1 file,  3,074 lines, 127 tests, 0 failing

Removed as redundant:

  * three copies of FakeLLM, four of the exporter and `spans` fixtures, and
    three or four copies each of `_attrs`, `_by_name` and `_all_by_name`
  * `break_agent` and `break_member_run`, which were the same function under
    two names, now one implementation plus a wrapper that keeps the second
    name's distinct default message
  * `test_all()` in test_user_utils.py, which re-called the other four tests
    *and* ran at module import, so those four executed twice per session

The `fake_agent` factories were not interchangeable - the advanced suite set
output_type="final" so structured replies survive the Agent round-trip - so
that is one implementation plus a `fake_agent_final` adapter.

Two bugs fixed along the way, both pre-existing:

  * The three failing tests were a stale test double. PlannerWorkerSwarm calls
    `self._run_judge(img=img)`, but the stub was `def fake_run_judge():` - the
    production signature gained `img` and the double was never updated.
  * `requires_llm` skipped only when NO provider key was set, but the live
    tests build an Agent on gpt-4o-mini. With a Groq or Gemini key present and
    no OpenAI key they ran anyway and failed on an empty credential. Now gated
    on OPENAI_API_KEY.

Also promotes the invariants suite's `_no_llm_keys` fixture to cover the whole
file, so an offline test cannot silently reach a real provider; the live class
opts out via a marker.

black and ruff clean.
